Refactor directory structure to move interfaces under Repositories
