+++ app/Http/Resources/NoteResource.php
@@ -18,9 +18,7 @@ class NoteResource extends JsonResource
             'id' => $this->id,
             'title' => $this->title,
             'description' => $this->description,
-            'category_id' => $this->category_id,
-            'created_at' => $this->created_at,
-            'updated_at' => $this->updated_at
+            'category_id' => $this->category_id
         ];
     }
 }
