+++ app/Repositories/Interfaces/RepositoryInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Interfaces;
+namespace App\Repositories\Interfaces;
 
 interface RepositoryInterface
 {
