+++ app/Services/NoteService.php
@@ -2,11 +2,9 @@
 
 namespace App\Services;
 
-use App\Models\Category;
-use App\Models\Note;
-use App\DTOs\NoteDTO;
-use App\Repositories\NoteRepository;
 use App\Interfaces\RepositoryInterface;
+use App\DTOs\NoteDTO;
+use App\Models\Note;
 
 class NoteService
 {
@@ -16,14 +14,6 @@ class NoteService
     {
         $this->noteRepository = $noteRepository;
     }
-
-    private function validateNoteCategory(Note $note, Category $category)
-    {
-        if ($note->category_id !== $category->id) {
-            throw new \Exception('Note does not belong to this category');
-        }
-    }
-    // General Methods (for all notes)
 
     public function getAll()
     {
@@ -52,32 +42,4 @@ class NoteService
     {
         return $this->noteRepository->delete($note);
     }
-
-    ////////////////////////////// Category-Specific Methods //////////////////////////////////////
-
-    public function createNoteInCategory(NoteDTO $dto, Category $category)
-    {
-        return $this->noteRepository->create([
-            'title' => $dto->title,
-            'description' => $dto->description,
-            'category_id' => $category->id,
-        ]);
-    }
-
-    public function updateNoteInCategory(Note $note, NoteDTO $dto, Category $category)
-    {
-        $this->validateNoteCategory($note, $category);
-        return $this->noteRepository->update($note, [
-            'title' => $dto->title,
-            'description' => $dto->description,
-            'category_id' => $category->id,
-        ]);
-    }
-
-    public function deleteNoteFromCategory(Category $category, Note $note)
-    {
-        $this->validateNoteCategory($note, $category);
-        return $this->noteRepository->delete($note);
-    }
 }
-
