+++ routes/api.php
@@ -4,6 +4,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\NoteController;
 use App\Http\Controllers\CategoryController;
+use App\Http\Controllers\CategoryNoteController;
 
 /*
 |--------------------------------------------------------------------------
@@ -20,5 +21,5 @@ Route::apiResource('categories', CategoryController::class);
 Route::apiResource('notes', NoteController::class);
 
 Route::prefix('categories')->group(function () {
-    Route::apiResource('{category}/notes', NoteController::class);
+    Route::apiResource('{category}/notes', CategoryNoteController::class);
 });
